Move extension tests into their own test module.
